% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{theme_cdl}
\alias{theme_cdl}
\title{CDL default ggpolot theme.}
\usage{
theme_cdl(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{fase_family}{base font family}
}
\value{
Return theme function for ggplots
}
\description{
CDL default ggpolot theme.
}
\examples{
ggplot() + theme_cdl()

}

